requirements.txt
kaggle
pandas
numpy
matplotlib
scikit-learn
seaborn
torch
torchvision
opencv-python
albumentations
Pillow
tqdm
git+https://github.com/huggingface/transformers.git
sam3
open_clip_torch
sentencepiece
numba
datasets
huggingface_hub
pyyaml
